+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Back, forward and refresh methods
   * Support for node 8.0.0
   * Major code refactor
+  * Mixwith dependency added
 
 0.4.6 / 2018-03-24
 ==================